requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
clip @ git+https://github.com/openai/CLIP.git
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Training and evaluation
wandb>=0.15.0
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0

# Data handling
dvc>=3.0.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
jupyter>=1.0.0